+++ cookbooks/myapp/recipes/openstack.rb
@@ -1,14 +1,10 @@
-include_recipe 'chef-metal'
-
-chef_gem 'fog'
-
-require 'chef_metal/fog'
+require 'chef_metal_fog'
 
 openstack_testdir = File.expand_path('~/openstack_test')
 
 directory openstack_testdir
 
-with_fog_provisioner :provider => 'OpenStack',
+with_fog_driver :provider => 'OpenStack',
     :openstack_api_key => ENV['OS_PASSWORD'],
     :openstack_username => ENV['OS_USERNAME'],
     :openstack_auth_url => ENV['OS_AUTH_URL'],
@@ -25,7 +21,7 @@ fog_key_pair 'me' do
   public_key_path "#{openstack_testdir}/me.pub"
 end
 
-with_provisioner_options :flavor_ref => ENV['OS_FLAVOR_REF']
+with_machine_options :flavor_ref => ENV['OS_FLAVOR_REF']
 
 machine 'chef_test' do
   recipe 'apt'